requirements.txt
requests
SQLAlchemy>=2.0
PyMySQL
python-whois
